Early config/server

+++ pokeleague.py
@@ -1,10 +1,19 @@
-import os, sys, shutil
-from modules import users
+import os, sys, shutil, asyncio
+from modules import users, config, server
 
 # boot script
 if '-clean' in sys.argv:
-    shutil.rmtree('_db')
+    print('This will wipe all data, including leagues and settings. Continue? (Y/N)')
+    res = input('> ')
+    if res.lower() == 'y':
+        shutil.rmtree('_db')
+        shutil.rmtree('_config')
 if not os.path.isdir('_db'):
     os.mkdir('_db')
+if not os.path.isdir('_config'):
+    os.mkdir('_config')
+# load modules
+config.init()
 users.init()
-users.register('TestGuy')
+# start server threads
+asyncio.run(server.start())
